Main method added in BubbleSort file
