+++ bubble_sort/Java/BubbleSort.java
@@ -4,6 +4,7 @@
 * */
 
 import java.util.ArrayList;
+import java.util.Scanner;
 
 public class BubbleSort
 {
@@ -54,5 +55,27 @@ public class BubbleSort
             System.out.print(arr.get(i) + " ");
         }
         System.out.println();
+    }
+
+    public static void main(String[] args)
+    {
+        Scanner scanner = new Scanner(System.in);
+
+        ArrayList<Integer> arr = new ArrayList<>();
+
+        int amount;
+        System.out.println("\nHow many numbers are there on which you want to apply bubble sort ?");
+        amount = scanner.nextInt();
+
+        int tempNo;
+        System.out.println("\nEnter numbers on which you want to apply bubble sort");
+        for (int i = amount - 1; i >= 0; i--)
+        {
+            tempNo = scanner.nextInt();
+            arr.add(tempNo);
+        }
+
+        BubbleSort bubbleSort = new BubbleSort(arr);
+        bubbleSort.printSortedList();
     }
 }
